style: Reformarteo el codigo se dividio lo background y agrego index
