+++ app/page.tsx
@@ -1,43 +1,22 @@
 import Image from "next/image";
-import {obtenerCommentarios, Paginacion} from "@/actions/comments";
-import {Comment, CommentProps} from "@/components/comment";
-import {Input} from "@/components/ui/input";
+import {obtenerCommentarios, Paginacion} from "@/actions";
+import {Comment, CommentProps, Input} from "@/components";
 import type {Metadata} from "next";
+import {BackgroundRadial} from "@/components/ui/BackgroundRadial";
+import {BackgroundLinear} from "@/components/ui/BackgroundLinear";
+import {BackgroundComment} from "@/components/comment/BackgroundComment";
 
 export const metadata: Metadata = {
     title: "Home",
     description: "Generated by create next app",
 }
 export default function Home() {
-    const commentsLeft: Paginacion<CommentProps> = obtenerCommentarios(0, 3);
-    const commentsMiddle: Paginacion<CommentProps> = obtenerCommentarios(1, 3);
-    const commentsRight: Paginacion<CommentProps> = obtenerCommentarios(2, 3);
     return (
         <main>
             <div className="relative w-full h-[calc(100vh-8rem)] overflow-hidden">
-                <div className="z-20 absolute -top-32 flex items-center justify-center w-full">
-                    <div className="grid gap-2 min-w-max lg:grid-cols-3 md:grid-cols-2 sm:grid-cols-1 grid-cols-1">
-                        <div className="grid gap-2">
-                            {commentsLeft.data.map((comments) => {
-                                return <Comment key={comments.id.toString()} {...comments}/>
-                            })}
-                        </div>
-                        <div className="grid gap-2 mt-32">
-                            {commentsMiddle.data.map((comments) => {
-                                return <Comment key={comments.id.toString()} {...comments}/>
-                            })}
-                        </div>
-                        <div className="grid gap-2">
-                            {commentsRight.data.map((comments) => {
-                                return <Comment key={comments.id.toString()} {...comments}/>
-                            })}
-                        </div>
-                    </div>
-                </div>
-                <div
-                    className="z-20 absolute top-0 w-full h-full bg-[radial-gradient(ellipse_at_center,_var(--tw-gradient-stops))] from-black from-0% to-transparent to-70% "></div>
-                <div className="z-20 absolute bottom-0 w-full h-96 bg-gradient-to-t from-black"></div>
-                <div className="z-20 absolute top-0 w-full h-32 bg-gradient-to-b from-black from-0%%"></div>
+                <BackgroundComment/>
+                <BackgroundRadial/>
+                <BackgroundLinear/>
                 <div className="absolute h-full flex items-center justify-center w-full mb-40 z-50">
                     <Image alt="logo" src="/static/images/logo.png" width="250" height="250"/>
                 </div>
