+++ components/index.ts
@@ -1,0 +1,4 @@
+export * from './comment/Comment'
+export * from './layout/Header'
+export * from './ui/avatar'
+export * from './ui/input'
