+++ components/ui/BackgroundLinear.tsx
@@ -1,0 +1,8 @@
+export function BackgroundLinear() {
+    return (
+        <>
+            <div
+                className="z-20 absolute top-0 w-full h-full bg-[radial-gradient(ellipse_at_center,_var(--tw-gradient-stops))] from-black from-0% to-transparent to-70% "></div>
+            <div className="z-20 absolute bottom-0 w-full h-96 bg-gradient-to-t from-black"></div>
+        </>)
+}
