+++ components/ui/BackgroundRadial.tsx
@@ -1,0 +1,5 @@
+export function BackgroundRadial() {
+    return (
+        <div
+            className="z-20 absolute top-0 w-full h-full bg-[radial-gradient(ellipse_at_center,_var(--tw-gradient-stops))] from-black from-0% to-transparent to-70% "/>)
+}
